fix(consensus): signed-byte header/block version comparisons (parity) (#106)
Follow-up to the header-serializer signed-version fix (#103): Scala models the
header/block version as a signed `Byte`, but several remaining comparisons read
it as unsigned `u8`. They agree for all real versions (1-4) and diverge only for
a malformed version > 127 (signed-negative) — computationally unreachable (such
a version cannot ride a PoW-valid header), but a genuine byte-grammar/parity gap
the differential fuzzer's header-version triage surfaced. Each is a
comparison-only one-liner (no arithmetic), so a signed cast is exact:

- ergo-ser block_transactions: the v2+ marker write gate (`block_version > 1`).
  Signed → no marker for version > 127, matching `BlockTransactionsSerializer`.
- ergo-sigma method_call: the `SGlobal.serialize(SHeader)` cost gate for the
  unparsed-bytes section (`h.version > INITIAL_VERSION`). Now matches the
  (signed) header writer from #103, so the serialize cost agrees with the bytes
  actually emitted.
- ergo-validation heights: the HardeningVersion gate (`block_version <= 2`,
  rule 124 monotonic heights).
- ergo-validation voting: the Interpreter60Version gate (`header.version >= 4`,
  rule 414 exMatchParameters60).

Deliberately NOT changed: `activated_script_version` (= `version - 1`). Byte-exact
parity there needs a signed (`i8`) representation across 129 references in 21
crates — disproportionate for an unreachable case — and a type-preserving clamp
is provably wrong because Scala's `Byte` arithmetic WRAPS (version 128 → `-128 -
1` wraps to `127`, not negative). Left as documented unreachable parity.

Test plan: cargo fmt --all -- --check; cargo clippy --workspace --all-targets
--all-features -- -D warnings; cargo test --workspace (all green). New
v2_marker_gate_is_signed_no_marker_above_version_127 pins the signed marker gate
(a v195 block serializes to the markerless v1 layout; a real v2 block keeps the
marker).

+++ ergo-sigma/src/evaluator/opcodes/method_call.rs
@@ -2883,7 +2883,13 @@ pub(in crate::evaluator) fn serialize_put_cost(
             c += chunk(4); // write_nbits — 4-byte compact difficulty
                            // put_u32(height) = putUInt = 0 (no charge)
             c += chunk(h.votes.len());
-            if h.version > ergo_ser::header::INITIAL_VERSION {
+            // Signed-Byte version comparison, matching the header writer
+            // (`ergo_ser::header::write_header_without_pow`) so this serialize
+            // cost agrees with the bytes actually emitted: a version > 127 is
+            // signed-negative, so no unparsed-bytes section is written or
+            // charged. (Unreachable: a script can only serialize real context
+            // headers, versions 1-4.)
+            if (h.version as i8) > ergo_ser::header::INITIAL_VERSION as i8 {
                 c += 1; // put_u8(unparsed_bytes.len())
                 c += chunk(h.unparsed_bytes.len());
             }
